+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2.19.1
 
+- `Improvements` - The [Cypress](https://www.cypress.io) was integrated as the end-to-end testing framework
 - `Fix` - The problem with destroy() method [#1380](https://github.com/codex-team/editor.js/issues/1380).
 - `Fix` - add getter keyword to `block.mergeable` method [#1415](https://github.com/codex-team/editor.js/issues/1415).
 - `Fix` — Fix problem with entering to Editor.js by Tab key [#1393](https://github.com/codex-team/editor.js/issues/1393)