+++ docs/CHANGELOG.md
@@ -10,7 +10,7 @@
 - `Fix` — Fix problem with entering to Editor.js by Tab key [#1393](https://github.com/codex-team/editor.js/issues/1393)
 - `Fix` - Sanitize pasted block data [#1396](https://github.com/codex-team/editor.js/issues/1396).
 - `Fix` - Unnecessary block creation after arrow navigation at last non-default block[#1414](https://github.com/codex-team/editor.js/issues/1414)
-
+- `New` - API for toggling Block Setting added [#1442](https://github.com/codex-team/editor.js/issues/1421). 
 
 ### 2.19